# Enable FEATURE_MULTICASTSTUB_AS_IL for Windows x86