… src/coreclr/clrdefinitions.cmake
… src/coreclr/vm/frames.cpp
… src/coreclr/vm/frames.h
… src/coreclr/vm/i386/stublinkerx86.cpp
… src/coreclr/vm/ilstubcache.cpp
… src/coreclr/vm/stubhelpers.h
… src/coreclr/vm/stublink.cpp
… src/coreclr/vm/stublink.h
… src/coreclr/vm/stubmgr.cpp