Closes #103958 .

Local benchmark result for `Action<int>` on x86:

| Method         | Job        | Toolchain                               | Mean      | Error     | StdDev    | Ratio |
|--------------- |----------- |---------------------------------------- |----------:|----------:|----------:|------:|
| SingleArg      | Job-OSFVND | \x86-3-TuneBranchPrediction\corerun.exe |  3.464 ns | 0.0127 ns | 0.0119 ns |  0.95 |
| SingleArg      | Job-KBQPJV | \x86-main\corerun.exe                   |  3.661 ns | 0.0222 ns | 0.0208 ns |  1.00 |
|                |            |                                         |           |           |           |       |
| ManyCast       | Job-OSFVND | \x86-3-TuneBranchPrediction\corerun.exe |  8.188 ns | 0.0944 ns | 0.0883 ns |  0.97 |
| ManyCast       | Job-KBQPJV | \x86-main\corerun.exe                   |  8.483 ns | 0.0888 ns | 0.0831 ns |  1.00 |
|                |            |                                         |           |           |           |       |
| ManyArg_RetFPU | Job-OSFVND | \x86-3-TuneBranchPrediction\corerun.exe |  5.566 ns | 0.0338 ns | 0.0316 ns |  0.96 |
| ManyArg_RetFPU | Job-KBQPJV | \x86-main\corerun.exe                   |  5.771 ns | 0.0212 ns | 0.0177 ns |  1.00 |