@huoyaoyuan 2024-06-29
@EgorBot -intel -amd -arm64

```csharp
using BenchmarkDotNet.Attributes;
using System;

namespace BenchmarkGround
{
    public struct GCStruct
    {
        public object a, b, c, d, e, f, h, g;
    }

    public class Bench
    {
        private readonly object obj = new object();

        private readonly Guid guid = Guid.NewGuid();

        private readonly GCStruct gcStruct = new GCStruct
        {
            a = new object(),
            g = new object()
        };

        private readonly Action<int> singleArg = (Action<int>)delegate { } + delegate { };

        private readonly Action<int> manyCast = (Action<int>)delegate { }
        +
            delegate { }
        + delegate { }
        + delegate { }
        + delegate { }
        + delegate { };

        private readonly Func<int, double, Guid, GCStruct, object, double> manyArg_RetFPU = (Func<int, double, Guid, GCStruct, object, double>)
            delegate { return 123.456; }
        + delegate { return 654.321; };

        [Benchmark]
        public void SingleArg() => singleArg(42);

        [Benchmark]
        public void ManyCast() => manyCast(42);

        [Benchmark]
        public double ManyArg_RetFPU() => manyArg_RetFPU(42, 123.0, guid, gcStruct, obj);
    }
}
```